10/10/2016

+++ niveau 2 exercice 23.c
@@ -5,7 +5,7 @@
 int main(int argc, char *argv[])
 {
  int resistance=0;
-float R1, R2, R3 ,Rser, Rder;
+double R1, R2, R3 ,Rser, Rder;
 
  // choix dérivation ou en série :
 printf("les resistance sont branche : \n1.en serie\n2.en derivation\n") ;
@@ -15,29 +15,29 @@ switch (resistance)
 {
 case 1:
     printf(" quel est la valeur en ohm de votre premiere resistance R1 ? ");
-    scanf("%f",&R1);
+    scanf("%lf",&R1);
     printf(" quel est la valeur en ohm de votre deuxieme resistance R2 ? ");
-    scanf("%f",&R2);
+    scanf("%lf",&R2);
     printf(" quel est la valeur en ohm de votre troisieme resistance R3 ? ");
-    scanf("%f",&R3);
+    scanf("%lf",&R3);
 // calcul R1+R2+R3
 
 Rser = (R1+R2+R3);
-    printf("votre resistance Rser est egale a %f ",Rser);
+    printf("votre resistance Rser est egale a %lf ",Rser);
 
 break;
 
 case 2:
     printf(" quel est la valeur en ohm de votre premiere resistance R1 ? ");
-    scanf("%f",&R1);
+    scanf("%lf",&R1);
     printf(" quel est la valeur en ohm de votre deuxieme resistance R2 ? ");
-    scanf("%f",&R2);
+    scanf("%lf",&R2);
     printf(" quel est la valeur en ohm de votre troisieme resistance R3 ? ");
-    scanf("%f",&R3);
+    scanf("%lf",&R3);
 // calcul (R1*R2*R3)/((R1*R2)+(R1*R3)+(R2*R3))
 
 Rder = (R1*R2*R3)/(R1*R2+R1*R3+R2*R3);
- printf("votre resistance Rder est egale a %f ",Rder);
+ printf("votre resistance Rder est egale a %lf ",Rder);
 break;
 
 }
